requirements.txt
requests
python-dotenv
lxml
html5lib
beautifulsoup4
